Restructure rest and rpc packages
Have all server packages under `servers/`. This change is being done in
preperation for using supervisor trees to manage GD2 listeners.

The following package changes have been done,
`rpc/server` -> `servers/peerrpc`
`rest` -> `servers/rest`

The move was done using the `gomvpkg` command.
