+++ commands/peers/getpeers.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/gluster/glusterd2/peer"
-	"github.com/gluster/glusterd2/rest"
+	"github.com/gluster/glusterd2/servers/rest"
 )
 
 func getPeersHandler(w http.ResponseWriter, r *http.Request) {
