+++ servers/peerrpc/listener.go
@@ -1,4 +1,4 @@
-package server
+package peerrpc
 
 import (
 	"net"
